refactor: 동화갤러리 리팩토링 및 댓글 UI 개선
